+++ src/context.tsx
@@ -3,7 +3,6 @@ import {BaseTheme} from 'types';
 
 export const ThemeContext = React.createContext({
   colors: {},
-  spacing: {},
 });
 
 export const ThemeProvider = ({